+++ CHANGELOG.md
@@ -381,3 +381,4 @@
 * updated net-creds and added libs to lib_install.sh
 * added cap_crack to network_crack.py for WPA/WPA2 handshake cracking
 * fixed increment usage in network_crack.py and added optimization suggestions
+* added zsteg to lib_install.sh